[lexical-list] Feature: export registerCheckList
